Key agreement components must account for both encryption and decryption
Terminology of "recipient" and "originator" are reversed in the 2 cases,
so instead use "public" and "private" since that doesn't really change.

For decryption case, need to pass in the recipient/private credential
via KeyAgreementParameters.

In the KeyAgreementCredential production this means that have to
populate the recipientCredential and originatorCredential properties
based on which case it is.