Collapse sequence of a conditional and its inverse into an if/else.
No reason to run the second block's tests on the case where the
backup file didn't originally exist and a test file was created
and then deleted.

It's suspected that on Windows the test file deletion in block 1 might
sometimes be in a partially-deleted state such that in block 2
File#exists() is true but either #canRead() or #canWrite() are false,
causing an unnecessary failure.