OSJ-307: Support Santuario's XMLParser interface
Default to global ParserPool, fallback to an internal one if global
registry not yet initialized and available.

Get maxPoolSize from Configuration Properties.

Adjust class name to be consistent with other Santuario-specific
classes, and move to the package with the other kids.

Add to ApacheXMLSecurityInitializer, with a config prop option to
disable. Default is enabled (for now).